fixing part 2

+++ src/main/kotlin/net/dongliu/apk/parser/struct/resource/TypeSpec.kt
@@ -9,7 +9,7 @@ class TypeSpec(header: TypeSpecHeader, entryFlags: LongArray, name: String?) {
     val id: Short
 
     init {
-        id = header.id
+        id = header.id.toShort()
         this.entryFlags = entryFlags
         this.name = name
     }
